requirements.txt
mysql-connector-python>=9.0.0
tabulate>=0.9.0
